Finished checkUser
Some finishes and polish was added to the checkUser function. It now checks username length, and has more specific character checks to ensure a directory can be made with the email. It also now checks that the email is not used by another account.

Users folder was made to store the user profiles. All profiles have been given a placeholder text file so that they are recognized by github.