Fixed Account Management Bugs
Fixed a bug in account deletion that caused errors that would end the program. The user struct now will be set to NULL strings and 0 ints to ensure no issues when calling free().

Fixed an error in login where invalid memory would be read if the email was entered wrong, crashing the program.