Created makeNotebook function
Created a function to create notebooks. Functions to test the memory storage of the notebooks has not yet been tested, as ways to view the memory have not yet been written. Notebook names and creation time are stored in the Notebook_List.txt file, and have their own text files made under the Notebooks folder. The order of notebooks should be sorted within Notebook_List, as I don't think you can reorder the way the directory is read.