Started Adding Sorts
Some existing errors were fixed.
Time is now printed in a readable format.
Sorting notebooks by time and alphabet has been created.
An attempt at a method to save a changed notebook order has been created, but is commented out due to a bug that results in it not printing anything, leaving the list blank.
Shifts, Swaps, and sorts for notes still need to be added.