Finished Sorting
Sorts have been finished for both notes and notebooks.
Notebooks still aren't saved to the list properly.
An issue in account deletion caused by the notebook list not being deleted has been fixed.
An issue has been found, with notes note loading when the notebook is selected. This may be due to the function just not being implemented yet, but has note been checked.
A current to do list has been added in comments at the bottom of the file.